part 4d

+++ controllers/blogs.js
@@ -1,33 +1,71 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
+
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
   
-blogsRouter.post('/', async (request, response) => {
-  if(typeof request.body.likes === 'undefined') 
-    request.body = { likes: 0, ... request.body }
+blogsRouter.post('/', async (request, response, next) => {
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
-  if(typeof request.body.title === 'undefined' &&
-     typeof request.body.url === 'undefined')
-  {
-    response.sendStatus(400)
-    return
+    const user = await User.findById(decodedToken.id)
+
+    var likes = 0
+
+    if(typeof request.body.likes !== 'undefined') 
+      likes = request.body.likes
+
+    if(typeof request.body.title === 'undefined' &&
+      typeof request.body.url === 'undefined')
+    {
+      response.sendStatus(400)
+      return
+    }
+
+    const blog = new Blog({
+      title: request.body.title,
+      likes: likes,
+      author: request.body.author,
+      url: request.body.url,
+      user: user._id
+    })
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog.toJSON())
+  } catch(exception) {
+    next(exception)
   }
-
-  const blog = new Blog(request.body)
-
-  const savedBlog = await blog.save()
-  response
-    .status(201)
-    .json(savedBlog.toJSON())
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() !== decodedToken.id.toString())
+    {
+      return response.status(401).json({ error: 'you can only remove your own entries'})
+    }
+
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } catch(exception) {
+    next(exception)
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
